\name{d3donut-package}
\alias{d3donut-package}
\alias{d3donut}
\docType{package}
\title{
\packageTitle{d3donut}
}
\description{
\packageDescription{d3donut}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{d3donut}
\packageIndices{d3donut}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{d3donut}

Maintainer: \packageMaintainer{d3donut}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
